pyproject.toml
[tool.poetry]
name = "bring-a-crew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ollama = "^0.4.7"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
